New test cases were transferred into TestBugFixes class
